Bug 356936 - Move DefaultCellRenderer to a none-internal package
